refactor: separate initial mount into its own folder
Move initial-mount-only code into `render/initial-mount/`:
- `build-node.ts` — buildNode() (from mount.ts)
- `mount-component.ts` — createComponentInstance(), mountComponent(),
  initialComponentRender() (merged from component/index.ts + component/initial.ts)

The `component/` folder now only contains rerender/resume code:
- `lifecycle.ts` — shared helpers (runComponentRender, commitRender, etc.)
- `rerender.ts` — executeComponentRerender, resumeInstance, rerenderInstance

Deleted: mount.ts, component/index.ts, component/initial.ts

Closes #169
